+++ grappelli/static/grappelli/js/grappelli.js
@@ -64,8 +64,8 @@ var django = {
         
         // HACK: adds an event listener to the today button of datepicker
         // if clicked today gets selected and datepicker hides.
-        // use live() because couldn't find hook after datepicker generates it's complete dom.
-        $(".ui-datepicker-current").live('click', function() {
+        // use on() because couldn't find hook after datepicker generates it's complete dom.
+        $(".ui-datepicker-current").on('click', function() {
             $.datepicker._selectDate(grappelli.datepicker_instance);
             grappelli.datepicker_instance = null;
         });
@@ -75,7 +75,7 @@ var django = {
 
         // now-button for both date and time
         // $("<button class='ui-datetime-now' />").insertAfter("button.ui-timepicker-trigger");
-        // $(".ui-datetime-now").live('click', function() {
+        // $(".ui-datetime-now").on('click', function() {
         //     alert("Now for date and time: grappelli.js line 68 ff.");
         //     return false
         // });