+++ grappelli/static/grappelli/js/jquery.grp_autocomplete_m2m.js
@@ -163,7 +163,7 @@
                     return false;
                 }
             })
-            .data("autocomplete")._renderItem = function(ul,item) {
+            .data("ui-autocomplete")._renderItem = function(ul,item) {
                 if (!item.value) {
                     return $("<li></li>")
                         .data( "item.autocomplete", item )